Fixing test cases.
